Refactor: create products package and move the screen to it
