+++ src/main/java/com/fp3/haras/view/container/FeatureWrapper.java
@@ -8,7 +8,7 @@ import com.fp3.haras.utils.Screens;
 import com.fp3.haras.view.screens.AnimalsScreen;
 import com.fp3.haras.view.screens.CustomersScreen;
 import com.fp3.haras.view.screens.home.HomeScreen;
-import com.fp3.haras.view.screens.ProductsScreen;
+import com.fp3.haras.view.screens.products.ProductsScreen;
 import com.fp3.haras.view.screens.StableScreen;
 
 public class FeatureWrapper extends javax.swing.JPanel implements GenericObserver<Screens> {
